ndvm/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.2
contourpy==1.1.0
cycler==0.11.0
distance-metrics==0.0.2
filelock==3.12.2
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
idna==3.4
imbalanced-learn==0.11.0
imblearn==0.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
lightning-utilities==0.9.0
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==2.0.7
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
torch==2.0.1
torchmetrics==1.0.3
tqdm==4.66.1
triton==2.0.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
xgboost==1.7.6
yarl==1.9.2
zipp==3.16.2
